successful token check
signin page now updates local storage token status and redirects to dashboard. dashboard redirects to signin if no token is found. rather than using props, I used the reactContext hooks createContext and useContext to track token status in the App.jsx, making it so it can be globally update with the hook in any child component.