rudimentary dashboard
get all rooms functions in nav component able to be imported to dashboard and specific room pages. next is adding functionality so that the list html elements will take you to that room's page on click. each room page should have a unique url and display messages in that room. (messages routes)